completed some order services
-  GetDeliveredOrdersByUserIdAsync
- GetUnDeliveredOrdersByUserIdAsync
- GetOrderDetailsByUserAsync
- AddOrderAsync
- UpdateOrderStatusAsync

+++ Manzili/backend/ManziliApi/Manzili.Core/Entities/OrderProduct.cs
@@ -7,9 +7,9 @@
         public int ProductId { get; set; } // FK
 
 
-        public int Price { get; set; }
+        public double Price { get; set; }
         public int Quantity { get; set; } = 1;
-        public int TotlaPrice { get; set; }
+        public double TotlaPrice { get; set; }
